+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+## 0.3.0
+
+- Adds pressure strokes (`perfect-freehand`).
+- Makes all strokes fade after a certain time has passed since the last stroke.
+
 ## 0.2.1
 
 - Notarized / signed mac app.
 
 ## 0.2.0
 
-- Adds changelog.
+- Adds change log.